pitchfork_trainer/requirements.txt
appdirs==1.4.4
black==21.4b2
boto3==1.17.62
botocore==1.20.62
click==7.1.2
jmespath==0.10.0
joblib==1.0.1
mypy-extensions==0.4.3
numpy==1.20.2
pandas==1.2.4
pathspec==0.8.1
pyathena==2.2.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
tenacity==7.0.0
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.4
